Move mock overrides to tests base setup
