--- __mocks__/openmrs-esm-framework.mock.tsx
@@ -1,13 +1,0 @@
-import { ConfigMock } from "./chart-widgets-config.mock";
-import { mockPatient } from "./patient.mock";
-export * from "@openmrs/esm-framework/mock";
-
-export const useCurrentPatient = jest.fn(() => {
-  return [false, mockPatient, mockPatient.id, null];
-});
-
-export function useConfig() {
-  return ConfigMock;
-}
-
-export const mockPatientId = mockPatient.id;
