+++ jest.config.json
@@ -8,8 +8,7 @@
     "\\.(jpg|jpeg|png)$": "<rootDir>/__mocks__/file.mock.ts",
     "^@carbon/icons-react/es/(.*)$": "@carbon/icons-react/lib/$1",
     "^carbon-components-react/es/(.*)$": "carbon-components-react/lib/$1",
-    "@openmrs/esm-framework/mock": "@openmrs/esm-framework/mock",
-    "@openmrs/esm-framework": "<rootDir>/__mocks__/openmrs-esm-framework.mock.tsx",
+    "@openmrs/esm-framework": "@openmrs/esm-framework/mock",
     "@carbon/charts/interfaces/enums":"<rootDir>/__mocks__/carbon-interface.ts",
     "^lodash-es/(.*)$": "lodash/$1"
   },
